getStandardTx
